Support Less 1.4.0
- Use builtin Less option to import synchronously
- Don't monkeypatch Less anymore

We had to instantiate a `libs.less.tree.parseEnv` object and set syncImport to true manually on it in order to circumvent a bug in Less (https://github.com/less/less.js/issues/1389) instead of passing a simple option objects. When the bug is fixed upstream, syncImport can be set again as a property of an option object passed to the Parser constructor.