Remove unneeded expectations from tests.
It doesn't matter that files may exist on the disk, because we always
overwrite them at the start of the tests to known values.

They shouldn't exist on disk anyways -- as they are cleaned up by tests -- but
they will exist on disk if a test fails and is then re-run.