I give up. Flip test assertion.
Unfortunately, I wasn't able to get Mincer to always update the manifest, even when an asset changes on disk.

Since this is an edge case (only happens when using manifests in development and changes are made), I won't hold up v3 for it. I'd love for it to change, though, and have logged that in comments in the test.