Include bin script in test coverage.
Changes bin script to execute only if not required as a module, and then includes it during the test run. This exposes the file to blanket, and so it’s included in our test coverage reports.

Needless to say, this commit fails the test coverage threshold (it’s down to 81%). New tests to come.