Touch up test logic on CDN assertions.
Instead of dynamically generating a host (and using it), we can just use a static host (as this is more like how this feature will be actually used). We also don’t need to actually retrieve that URL at all during the test, since we’re only concerned that the URL is correctly generated.